* commit '7ce6e78563893ef8da687787bd96d9b7d4af1ace':
  Fix a long subId.

+++ src/com/android/server/telecom/PhoneAccountRegistrar.java
@@ -211,7 +211,7 @@ public final class PhoneAccountRegistrar {
                     PhoneAccount.CAPABILITY_SIM_SUBSCRIPTION)) {
                 // If the account selected is a SIM account, propagate down to the subscription
                 // record.
-                long subId = getSubscriptionIdForPhoneAccount(accountHandle);
+                int subId = getSubscriptionIdForPhoneAccount(accountHandle);
                 SubscriptionManager.setDefaultVoiceSubId(subId);
             }
